refactor(metrics): reorganize files
